calculo de aranceles

+++ Sistema-Matricula-3/estudiante.h
@@ -19,12 +19,11 @@ class estudiante :public objeto{
 public:
     estudiante(string,string,string,string,string,string,float);
     virtual ~estudiante();
-    virtual double calcularPagoCreditos() = 0;
+    virtual long calcularPagoCreditos() = 0;
     virtual string toString()const ;
     
     virtual string obtenerNombre()const;
-//    virtual string obtenerApellido1()const;
-//    virtual string obtenerApellido2()const;
+
     virtual string obtenerCedula()const;
     virtual string obtenerCarnet()const;
     virtual string obtenerNacionalidad()const;
@@ -33,12 +32,18 @@ public:
     virtual void asignarNombre(string);
     virtual void asignarApellido1(string);
     virtual void asignarApellido2(string);
-//    virtual void asignarCedula(string);
-//    virtual void asignarCarnet(string);
     virtual void asignarNacionalidad(string);
     virtual void asignarPorcentajeBeca(float);
     
     virtual void asignarCurso(objeto*) = 0;
+    
+    
+    
+//    virtual void asignarCedula(string);
+//    virtual void asignarCarnet(string);
+//    virtual string obtenerApellido1()const;
+//    virtual string obtenerApellido2()const;
+    
 protected:
     string _nombre;
     string _apellido1;
